plugins/ui/web/golismero_webapp/requirements.txt
django
djangorestframework
markdown
django-filter
pygments
django-cors-headers